backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0
